Reduce Code Duplication in Food Card
Make reusable components out of shared code.
